+++ hive/example/lib/hive/hive_registrar.g.dart
@@ -2,12 +2,11 @@
 // Do not modify
 // Check in to version control
 
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:example/hive/hive_adapters.dart';
 
 extension HiveRegistrar on HiveInterface {
   void registerAdapters() {
-    registerAdapter(FreezedPersonAdapter());
     registerAdapter(JobAdapter());
     registerAdapter(PersonAdapter());
   }
@@ -15,7 +14,6 @@ extension HiveRegistrar on HiveInterface {
 
 extension IsolatedHiveRegistrar on IsolatedHiveInterface {
   void registerAdapters() {
-    registerAdapter(FreezedPersonAdapter());
     registerAdapter(JobAdapter());
     registerAdapter(PersonAdapter());
   }