+++ hive/example/lib/inspector_example.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:example/hive/hive_registrar.g.dart';
 import 'package:example/main.dart';
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 
 void main() async {
   final path = Directory.current.path;
@@ -11,7 +11,6 @@ void main() async {
   await IsolatedHive.init(path);
   IsolatedHive.registerAdapters();
 
-  final box1 = await Hive.openBox('testBox');
   final box2 = await Hive.openBox('testBox2');
 
   final john = Person(name: 'John', age: 30);
@@ -44,8 +43,26 @@ void main() async {
     await box5.put(i, john);
   }
 
+  bump();
+  toggleBoxRegistration();
+}
+
+void bump() async {
+  final box1 = await Hive.openBox('testBox');
   while (true) {
     await Future.delayed(const Duration(seconds: 1));
+    print('bump');
     await box1.add('bump');
+  }
+}
+
+void toggleBoxRegistration() async {
+  while (true) {
+    final box = await Hive.openBox('tempBox');
+    print('tempBox opened');
+    await Future.delayed(const Duration(seconds: 5));
+    await box.close();
+    print('tempBox closed');
+    await Future.delayed(const Duration(seconds: 5));
   }
 }