+++ hive/lib/src/backend/vm/backend_manager.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce/src/backend/storage_backend.dart';
 import 'package:hive_ce/src/backend/vm/storage_backend_vm.dart';
 import 'package:hive_ce/src/util/obfuscation_utils.dart';
@@ -26,6 +26,7 @@ class BackendManager implements BackendManagerInterface {
     String? path,
     bool crashRecovery,
     HiveCipher? cipher,
+    int? keyCrc,
     String? collection,
     bool obfuscateBoxNames,
   ) async {
@@ -49,7 +50,8 @@ class BackendManager implements BackendManagerInterface {
     final (file, lockFile) =
         await findHiveFileAndCleanUp(name, path, obfuscateBoxNames);
 
-    final backend = StorageBackendVm(file, lockFile, crashRecovery, cipher);
+    final backend =
+        StorageBackendVm(file, lockFile, crashRecovery, cipher, keyCrc);
     await backend.open();
     return backend;
   }