+++ hive/lib/src/binary/frame_helper.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce/src/binary/binary_reader_impl.dart';
 import 'package:hive_ce/src/box/keystore.dart';
 
@@ -11,7 +11,8 @@ class FrameHelper {
     Uint8List bytes,
     Keystore? keystore,
     TypeRegistry registry,
-    HiveCipher? cipher, {
+    HiveCipher? cipher,
+    int? keyCrc, {
     bool verbatim = false,
   }) {
     final reader = BinaryReaderImpl(bytes, registry);
@@ -21,6 +22,7 @@ class FrameHelper {
 
       final frame = reader.readFrame(
         cipher: cipher,
+        keyCrc: keyCrc,
         lazy: false,
         frameOffset: frameOffset,
         verbatim: verbatim,