+++ hive/lib/src/isolate/handler/isolated_hive_handler.dart
@@ -1,8 +1,9 @@
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce/src/box/default_compaction_strategy.dart';
 import 'package:hive_ce/src/box/default_key_comparator.dart';
 import 'package:hive_ce/src/hive_impl.dart';
 import 'package:hive_ce/src/isolate/handler/isolated_box_handler.dart';
+import 'package:hive_ce/src/util/logger.dart';
 import 'package:isolate_channel/isolate_channel.dart';
 
 /// Method call handler for Hive methods
@@ -18,42 +19,56 @@ Future<dynamic> handleHiveMethodCall(
         obfuscateBoxNames: call.arguments['obfuscateBoxNames'] ?? false,
       );
       (Hive as HiveImpl).setIsolated();
+      final loggerLevel = call.arguments['logger_level'];
+      Logger.level = LoggerLevel.values.byName(loggerLevel);
     case 'openBox':
       final name = call.arguments['name'];
+      final lazy = call.arguments['lazy'];
+
       if (boxHandlers.containsKey(name)) {
         // Ensure this is a valid `openBox` call
-        Hive.box(name);
+        if (lazy) {
+          Hive.lazyBox(name);
+        } else {
+          Hive.box(name);
+        }
         return;
       }
 
-      final box = await Hive.openBox(
-        name,
-        keyComparator: call.arguments['keyComparator'] ?? defaultKeyComparator,
-        compactionStrategy:
-            call.arguments['compactionStrategy'] ?? defaultCompactionStrategy,
-        crashRecovery: call.arguments['crashRecovery'],
-        path: call.arguments['path'],
-        bytes: call.arguments['bytes'],
-        collection: call.arguments['collection'],
-      );
-      boxHandlers[name] = IsolatedBoxHandler(box, connection);
-    case 'openLazyBox':
-      final name = call.arguments['name'];
-      if (boxHandlers.containsKey(name)) {
-        // Ensure this is a valid `openLazyBox` call
-        Hive.lazyBox(name);
-        return;
+      final keyCrc = call.arguments['keyCrc'];
+      final keyComparator =
+          call.arguments['keyComparator'] ?? defaultKeyComparator;
+      final compactionStrategy =
+          call.arguments['compactionStrategy'] ?? defaultCompactionStrategy;
+      final crashRecovery = call.arguments['crashRecovery'];
+      final path = call.arguments['path'];
+      final bytes = call.arguments['bytes'];
+      final collection = call.arguments['collection'];
+
+      final BoxBase box;
+      if (lazy) {
+        box = await (Hive as HiveImpl).openLazyBox(
+          name,
+          keyCrc: keyCrc,
+          keyComparator: keyComparator,
+          compactionStrategy: compactionStrategy,
+          crashRecovery: crashRecovery,
+          path: path,
+          collection: collection,
+        );
+      } else {
+        box = await (Hive as HiveImpl).openBox(
+          name,
+          keyCrc: keyCrc,
+          keyComparator: keyComparator,
+          compactionStrategy: compactionStrategy,
+          crashRecovery: crashRecovery,
+          path: path,
+          bytes: bytes,
+          collection: collection,
+        );
       }
 
-      final box = await Hive.openLazyBox(
-        name,
-        keyComparator: call.arguments['keyComparator'] ?? defaultKeyComparator,
-        compactionStrategy:
-            call.arguments['compactionStrategy'] ?? defaultCompactionStrategy,
-        crashRecovery: call.arguments['crashRecovery'],
-        path: call.arguments['path'],
-        collection: call.arguments['collection'],
-      );
       boxHandlers[name] = IsolatedBoxHandler(box, connection);
     case 'deleteBoxFromDisk':
       await Hive.deleteBoxFromDisk(