+++ hive/lib/src/isolate/isolated_hive_impl/impl/isolated_hive_impl_vm.dart
@@ -2,14 +2,14 @@ import 'dart:async';
 import 'dart:isolate';
 import 'dart:typed_data';
 
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce/src/connect/hive_connect.dart';
 import 'package:hive_ce/src/isolate/handler/isolate_entry_point.dart';
 import 'package:hive_ce/src/isolate/isolated_box_impl/isolated_box_impl_vm.dart';
 import 'package:hive_ce/src/isolate/isolated_hive_impl/hive_isolate.dart';
 import 'package:hive_ce/src/isolate/isolated_hive_impl/hive_isolate_name.dart';
 import 'package:hive_ce/src/registry/type_registry_impl.dart';
-import 'package:hive_ce/src/util/debug_utils.dart';
+import 'package:hive_ce/src/util/logger.dart';
 import 'package:hive_ce/src/util/type_utils.dart';
 import 'package:isolate_channel/isolate_channel.dart';
 
@@ -29,13 +29,13 @@ class IsolatedHiveImpl extends TypeRegistryImpl
   @override
   IsolateConnection get connection => _connection!;
 
-  late Future<IsolateConnection> Function() _spawnHiveIsolate = () =>
-      spawnIsolate(
-        isolateEntryPoint,
-        debugName: hiveIsolateName,
-        onConnect: onConnect,
-        onExit: onExit,
-      );
+  late Future<IsolateConnection> Function() _spawnHiveIsolate =
+      () => spawnIsolate(
+            isolateEntryPoint,
+            debugName: hiveIsolateName,
+            onConnect: onConnect,
+            onExit: onExit,
+          );
 
   @override
   void onConnect(SendPort send) =>
@@ -57,8 +57,8 @@ class IsolatedHiveImpl extends TypeRegistryImpl
     if (_connection == null) {
       _isolateNameServer = isolateNameServer;
 
-      if (_isolateNameServer == null) {
-        debugPrint(HiveIsolate.noIsolateNameServerWarning);
+      if (Logger.noIsolateNameServerWarning && _isolateNameServer == null) {
+        Logger.w(HiveWarning.noIsolateNameServer);
       }
 
       final send =
@@ -66,7 +66,7 @@ class IsolatedHiveImpl extends TypeRegistryImpl
 
       final IsolateConnection connection;
       if (send != null) {
-        connection = connectToIsolate(send);
+        connection = await connectToIsolate(send);
       } else {
         connection = await _spawnHiveIsolate();
       }
@@ -76,10 +76,14 @@ class IsolatedHiveImpl extends TypeRegistryImpl
       _boxChannel = IsolateMethodChannel('box', connection);
     }
 
-    return _hiveChannel.invokeMethod('init', {
-      'path': path,
-      'obfuscateBoxNames': obfuscateBoxNames,
-    });
+    return _hiveChannel.invokeMethod(
+      'init',
+      {
+        'path': path,
+        'obfuscateBoxNames': obfuscateBoxNames,
+        'logger_level': Logger.level.name
+      },
+    );
   }
 
   Future<IsolatedBoxBase<E>> _openBox<E>(
@@ -123,6 +127,8 @@ class IsolatedHiveImpl extends TypeRegistryImpl
       try {
         final params = {
           'name': name,
+          'lazy': lazy,
+          'keyCrc': cipher?.calculateKeyCrc(),
           'keyComparator': comparator,
           'compactionStrategy': compaction,
           'crashRecovery': recovery,
@@ -131,26 +137,23 @@ class IsolatedHiveImpl extends TypeRegistryImpl
           'collection': collection,
         };
 
-        final IsolatedBoxBaseImpl<E> newBox;
-        if (lazy) {
-          await _hiveChannel.invokeMethod('openLazyBox', params);
-          newBox = IsolatedLazyBoxImpl<E>(
-            this,
-            name,
-            cipher,
-            connection,
-            _boxChannel,
-          );
-        } else {
-          await _hiveChannel.invokeMethod('openBox', params);
-          newBox = IsolatedBoxImpl<E>(
-            this,
-            name,
-            cipher,
-            connection,
-            _boxChannel,
-          );
-        }
+        await _hiveChannel.invokeMethod('openBox', params);
+
+        final newBox = lazy
+            ? IsolatedLazyBoxImpl<E>(
+                this,
+                name,
+                cipher,
+                connection,
+                _boxChannel,
+              )
+            : IsolatedBoxImpl<E>(
+                this,
+                name,
+                cipher,
+                connection,
+                _boxChannel,
+              );
 
         _boxes[name] = newBox;
 
@@ -180,17 +183,16 @@ class IsolatedHiveImpl extends TypeRegistryImpl
     String? collection,
   }) async =>
       await _openBox<E>(
-            name,
-            false,
-            encryptionCipher,
-            keyComparator,
-            compactionStrategy,
-            crashRecovery,
-            path,
-            bytes,
-            collection,
-          )
-          as IsolatedBox<E>;
+        name,
+        false,
+        encryptionCipher,
+        keyComparator,
+        compactionStrategy,
+        crashRecovery,
+        path,
+        bytes,
+        collection,
+      ) as IsolatedBox<E>;
 
   @override
   Future<IsolatedLazyBox<E>> openLazyBox<E>(
@@ -203,17 +205,16 @@ class IsolatedHiveImpl extends TypeRegistryImpl
     String? collection,
   }) async =>
       await _openBox<E>(
-            name,
-            true,
-            encryptionCipher,
-            keyComparator,
-            compactionStrategy,
-            crashRecovery,
-            path,
-            null,
-            collection,
-          )
-          as IsolatedLazyBox<E>;
+        name,
+        true,
+        encryptionCipher,
+        keyComparator,
+        compactionStrategy,
+        crashRecovery,
+        path,
+        null,
+        collection,
+      ) as IsolatedLazyBox<E>;
 
   IsolatedBoxBase<E> _getBoxInternal<E>(String name, bool lazy) {
     final lowerCaseName = name.toLowerCase();