+++ hive/test/integration/isolate_test.dart
@@ -4,12 +4,12 @@ library;
 import 'dart:async';
 import 'dart:isolate';
 
-import 'package:hive_ce/hive.dart';
-import 'package:hive_ce/src/backend/vm/storage_backend_vm.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce/src/hive_impl.dart';
 import 'package:hive_ce/src/isolate/handler/isolate_entry_point.dart';
 import 'package:hive_ce/src/isolate/isolated_hive_impl/hive_isolate.dart';
 import 'package:hive_ce/src/isolate/isolated_hive_impl/isolated_hive_impl.dart';
+import 'package:hive_ce/src/util/logger.dart';
 import 'package:isolate_channel/isolate_channel.dart';
 import 'package:test/test.dart';
 
@@ -122,7 +122,7 @@ void main() {
       group('warnings', () {
         test('unsafe isolate', () async {
           final patchedWarning =
-              HiveImpl.unsafeIsolateWarning.replaceFirst(isolateNameRegex, '');
+              HiveWarning.unsafeIsolate.replaceFirst(isolateNameRegex, '');
 
           final safeOutput = await Isolate.run(
             debugName: 'main',
@@ -137,6 +137,14 @@ void main() {
             unsafeOutput.first.replaceFirst(isolateNameRegex, ''),
             patchedWarning,
           );
+
+          final ignoredOutput = await Isolate.run(
+            () => captureOutput(() {
+              HiveLogger.unsafeIsolateWarning = false;
+              Hive.init(null);
+            }).toList(),
+          );
+          expect(ignoredOutput, isEmpty);
         });
 
         test('safe hive isolate', () async {
@@ -164,13 +172,19 @@ void main() {
               await captureOutput(() => IsolatedHiveImpl().init(null)).toList();
           expect(
             unsafeOutput,
-            contains(HiveIsolate.noIsolateNameServerWarning),
+            contains(HiveWarning.noIsolateNameServer),
           );
 
           final safeOutput = await captureOutput(
             () => IsolatedHiveImpl().init(null, isolateNameServer: TestIns()),
           ).toList();
           expect(safeOutput, isEmpty);
+
+          final ignoredOutput = await captureOutput(() {
+            HiveLogger.noIsolateNameServerWarning = false;
+            IsolatedHiveImpl().init(null);
+          }).toList();
+          expect(ignoredOutput, isEmpty);
         });
 
         test('unmatched isolation', () async {
@@ -180,18 +194,29 @@ void main() {
           await IsolatedHive.init(path, isolateNameServer: StubIns());
           Hive.init(path);
 
-          await IsolatedHive.openBox('test');
+          await IsolatedHive.openBox('box1');
           final output =
-              await captureOutput(() => Hive.openBox('test')).toList();
+              await captureOutput(() => Hive.openBox('box1')).toList();
 
           expect(
             output,
-            contains(StorageBackendVm.unmatchedIsolationWarning),
+            contains(HiveWarning.unmatchedIsolation),
+          );
+
+          await IsolatedHive.openBox('box2');
+          final ignoredOutput = await captureOutput(() async {
+            HiveLogger.unmatchedIsolationWarning = false;
+            await Hive.openBox('box2');
+          }).toList();
+
+          expect(
+            ignoredOutput,
+            isNot(contains(HiveWarning.unmatchedIsolation)),
           );
         });
       });
 
-      test('IsolatedHive data compatable with Hive', () async {
+      test('IsolatedHive data compatible with Hive', () async {
         final dir = await getTempDir();
 
         final isolatedHive = IsolatedHiveImpl();
@@ -209,9 +234,149 @@ void main() {
         final box = await hive.openBox('test');
         expect(await box.get('key'), 'value');
       });
+
+      test('Hive data compatible with IsolatedHive', () async {
+        final dir = await getTempDir();
+
+        final hive = HiveImpl();
+        addTearDown(hive.close);
+        hive.init(dir.path);
+
+        final box = await hive.openBox('test');
+        await box.put('key', 'value');
+        await box.close();
+
+        final isolatedHive = IsolatedHiveImpl();
+        addTearDown(isolatedHive.close);
+        await isolatedHive.init(dir.path, isolateNameServer: StubIns());
+
+        final isolatedBox = await isolatedHive.openBox('test');
+        expect(await isolatedBox.get('key'), 'value');
+      });
+
+      test('Encrypted IsolatedHive data compatible with Hive', () async {
+        final dir = await getTempDir();
+        final cipher = HiveAesCipher(Hive.generateSecureKey());
+
+        final isolatedHive = IsolatedHiveImpl();
+        addTearDown(isolatedHive.close);
+        await isolatedHive.init(dir.path, isolateNameServer: StubIns());
+
+        final isolatedBox =
+            await isolatedHive.openBox('test', encryptionCipher: cipher);
+        await isolatedBox.put('key', 'value');
+        await isolatedBox.close();
+
+        final hive = HiveImpl();
+        addTearDown(hive.close);
+        hive.init(dir.path);
+
+        final box = await hive.openBox('test', encryptionCipher: cipher);
+        expect(await box.get('key'), 'value');
+      });
+
+      test('Encrypted Hive data compatible with IsolatedHive', () async {
+        final dir = await getTempDir();
+        final cipher = HiveAesCipher(Hive.generateSecureKey());
+
+        final hive = HiveImpl();
+        addTearDown(hive.close);
+        hive.init(dir.path);
+
+        final box = await hive.openBox('test', encryptionCipher: cipher);
+        await box.put('key', 'value');
+        await box.close();
+
+        final isolatedHive = IsolatedHiveImpl();
+        addTearDown(isolatedHive.close);
+        await isolatedHive.init(dir.path, isolateNameServer: StubIns());
+
+        final isolatedBox =
+            await isolatedHive.openBox('test', encryptionCipher: cipher);
+        expect(await isolatedBox.get('key'), 'value');
+      });
+
+      test(
+        'IsolatedHive data encrypted with no keyCrc is readable',
+        () async {
+          final dir = await getTempDir();
+          final key = Hive.generateSecureKey();
+
+          final isolatedHive = IsolatedHiveImpl();
+          addTearDown(isolatedHive.close);
+          await isolatedHive.init(dir.path, isolateNameServer: StubIns());
+
+          final box = await isolatedHive.openBox(
+            'test',
+            encryptionCipher: ZeroKeyCrcCipher(key),
+          );
+          await box.put('key', 'value');
+          await box.close();
+
+          final box2 = await isolatedHive.openBox(
+            'test',
+            encryptionCipher: HiveAesCipher(key),
+          );
+          expect(await box2.get('key'), 'value');
+        },
+      );
+
+      test('Encrypted IsolatedBox data compatible with IsolatedLazyBox',
+          () async {
+        final dir = await getTempDir();
+        final key = Hive.generateSecureKey();
+
+        final isolatedHive = IsolatedHiveImpl();
+        addTearDown(isolatedHive.close);
+        await isolatedHive.init(dir.path, isolateNameServer: StubIns());
+
+        final box = await isolatedHive.openBox(
+          'test',
+          encryptionCipher: HiveAesCipher(key),
+        );
+        await box.put('key', 'value');
+        await box.close();
+
+        final lazyBox = await isolatedHive.openLazyBox(
+          'test',
+          encryptionCipher: HiveAesCipher(key),
+        );
+        expect(await lazyBox.get('key'), 'value');
+      });
+
+      test('Encrypted IsolatedLazyBox data compatible with IsolatedBox',
+          () async {
+        final dir = await getTempDir();
+        final key = Hive.generateSecureKey();
+
+        final isolatedHive = IsolatedHiveImpl();
+        addTearDown(isolatedHive.close);
+        await isolatedHive.init(dir.path, isolateNameServer: StubIns());
+
+        final lazyBox = await isolatedHive.openLazyBox(
+          'test',
+          encryptionCipher: HiveAesCipher(key),
+        );
+        await lazyBox.put('key', 'value');
+        await lazyBox.close();
+
+        final box = await isolatedHive.openBox(
+          'test',
+          encryptionCipher: HiveAesCipher(key),
+        );
+        expect(await box.get('key'), 'value');
+      });
     },
     onPlatform: {
       'chrome': Skip('Isolates are not supported on web'),
     },
   );
+}
+
+/// Test cipher that always returns a zero key CRC
+class ZeroKeyCrcCipher extends HiveAesCipher {
+  ZeroKeyCrcCipher(super.key);
+
+  @override
+  int calculateKeyCrc() => 0;
 }