+++ hive/test/tests/box_collection/box_collection_test.dart
@@ -1,4 +1,4 @@
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce/src/hive_impl.dart';
 import 'package:test/test.dart';
 
@@ -154,6 +154,30 @@ void main() {
       final box = await collection.openBox('cats', fromJson: Test.fromJson);
       await box.put('json_test', testObject);
       expect(await box.get('json_test'), testObject);
+    });
+
+    test('primitives', () async {
+      final collection = await _openCollection();
+      final box = await collection.openBox('cats');
+
+      const value = [
+        null,
+        1,
+        2,
+        [
+          3,
+          4,
+          [5, 6, null],
+        ],
+        {
+          'a': 1,
+          'b': [2, 3],
+          'c': {'d': 4, 'e': null},
+        },
+      ];
+
+      await box.put('nested_test', value);
+      expect(await box.get('nested_test'), value);
     });
   });
 }