+++ hive_generator/lib/src/adapter_builder/class_adapter_builder.dart
@@ -1,10 +1,13 @@
+// TODO: Remove with Dart 3.11
+// ignore_for_file: unnecessary_ignore, experimental_member_use
+
 import 'dart:typed_data';
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:collection/collection.dart';
-import 'package:hive_ce/hive.dart';
+import 'package:hive_ce/hive_ce.dart';
 import 'package:hive_ce_generator/src/adapter_builder/adapter_builder.dart';
 import 'package:hive_ce_generator/src/helper/helper.dart';
 import 'package:source_gen/source_gen.dart';