+++ hive_generator/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.11.1
+
+- Upgrades `analyzer` to `10.0.0`
+
+## 1.11.0
+
+- Adds support for `ignoredFields` in the schema migrator
+
+## 1.10.1
+
+- Upgrades `analyzer` to `9.0.0`
+
 ## 1.10.0
 
 - Adds support for `ignoredFields` parameter in `AdapterSpec` (by [@rezam92](https://github.com/rezam92) in [#229](https://github.com/IO-Design-Team/hive_ce/pull/229))